restructured commander and node functions
