+++ Commanders/Functions_Commander.py
@@ -7,12 +7,22 @@ def create_commanders (commanders_starting_data):
         commanders[indiv_cmdr.human] = indiv_cmdr
     return commanders
 
-def retrieve_cmdr_strings(human, commanders_starting_data):
-    for each_key in commanders_starting_data:
-        if each_key == human:
-            members = commanders_starting_data[each_key]["Unit Members"]
-            dots_owned = commanders_starting_data[each_key]["Dots Owned"]
-            country = commanders_starting_data[each_key]["Country"]
-            break
-    return members, dots_owned, country
-
+def update_commanders(commanders, nodes, commanders_data, units_data):
+    units = {}
+    for id in commanders:
+        commander = commanders[id]
+        # get class properties from data
+        for each_key in commanders_data:
+            if each_key == commander.human:
+                members = commanders_data[each_key]["Unit Members"]
+                dots_owned = commanders_data[each_key]["Dots Owned"]
+                country = commanders_data[each_key]["Country"]
+        # assign class properties
+        commander.assign_country(country)
+        commander.add_units(units_data, members, nodes)
+        # update units
+        units = {**units, **commander.unit_members}
+        # assign class properties
+        commander.retrieve_dots_owned(dots_owned, nodes)
+        commander.retrieve_hsc(dots_owned, nodes)
+    return commanders, units
